Move airportConditions from the plan object to the scenario object Fixes #145
